Refactor Suite: move .markedPending to result.
Back in 2.4 it() functions inside xdescribe() were status pending. After commit c21bdaf
these functions have status 'disabled' and we can now longer account for pending specs
separate from disabled ones.  By moving the markedPending onto the results, reporters
can determine if a spec is within a pending suite and properly count the spec as pending,
not as disabled (skipped).

Adding this property to the result should not have signficant effects; in fact it seems
like the pending mark should be part of the result sent to reporters.