Pass error arguments to error handlers.
Passing only the first argument cripples jasmine in web browser uses.  In the browser the onerror event is passed a funky list of args:
https://developer.mozilla.org/en-US/docs/Web/API/GlobalEventHandlers/onerror
By slicing only the first arg only get the message, never the lines and most important never the Error object with the call stack.

AFAIK in node the Error object is the sole object sent to the handler so we are no worse off in that case.